+++ bonus/dda.c
@@ -6,11 +6,11 @@
 /*   By: recherra <recherra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/25 20:50:48 by recherra          #+#    #+#             */
-/*   Updated: 2025/03/26 18:27:51 by recherra         ###   ########.fr       */
+/*   Updated: 2025/03/27 22:27:02 by recherra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../cube_bonus.h"
+#include "cube_bonus.h"
 
 static void	process_data(t_dda_data *dda, t_coords *start, t_coords *end)
 {
@@ -48,7 +48,7 @@ void	dda(t_all_data *data, int padding)
 {
 	t_coords	start;
 	t_coords	end;
-	double		ray_angle;
+	float		ray_angle;
 
 	start.x = data->minimap.tile * padding + 2;
 	start.y = start.x;
